Added logic that will allow user to edit existing orders (using AddOrder
window). The current selected table row in the table on MainWindow will
be used to populate the combos/textfields on AddOrder with the existing
order information. The add/edit mode is controlled by a class variable
in AddOrder that is set to the string "edit." This variable will only be
set to "edit" if AddOrder is launched from MainWindow using the "Edit
Order" button.